+++ CHANGELOG.md
@@ -6,6 +6,28 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.9.0] - 2023-03-27
+
+### Added
+
+-   `util rtt` command now supports `--logfile` with which you can specify file
+    into which to store RTT logs (#42).
+-   `debug` command which uses `west debug` or (`west attach`) to connect to the
+    board and start a debugging session.
+
+### Changed
+
+-   Add back functionality to provide `cmake_args` after `--` marker for `build`
+    command. Now it is possible to provide custom `-D` define values to the
+    CMake. Commands `flash`, `bypass` already provided option for extra
+    arguments after `--`, however they did it incorrectly for very edge cases,
+    as they removed double quotes from all arguments passed after `--` marker
+    (#56).
+
+### Fixed
+
+-   Ton of spelling mistakes in the code and comments.
+
 ## [0.8.0] - 2023-03-20
 
 ### Added
@@ -167,7 +189,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 -   Docker scripts for building and running docker containers, for development
     purposes.
 
-[Unreleased]: https://github.com/IRNAS/irnas-east-software/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/IRNAS/irnas-east-software/compare/v0.9.0...HEAD
+
+[0.9.0]: https://github.com/IRNAS/irnas-east-software/compare/v0.8.0...v0.9.0
 
 [0.8.0]: https://github.com/IRNAS/irnas-east-software/compare/v0.7.0...v0.8.0